feat: complete refactor and componentisation of code
Added state to the app to allow for using local changes. Connected
app to the API to allow for live updates of the shelf on which
each book sits. Each book sits on either the Reading, Read, or
want to read shelves. If none shelf is selected it will not be
displayed.

Plus button in the bottom right was modified to take the user to a
new page, /search, to search and add new books. All seach functions
removed and put into new Search component.

Books state was implemented as an object instead of a list. List
implementation creates the behaviour where th book jumps to the end
of the current shelf before moving to the new shelf. Also, object
implementaion keeps a consistent order even when a book is restored
to a previous shelf.